requirements_GH_actions.txt
joblib == 1.2.0
numpy == 1.23.5
pandas == 1.5.3
pyarrow == 12.0.0
pytest == 7.3.1
scikit-learn == 1.3.0
python-box==7.0.1
pyyaml == 6.0
bs4 == 0.0.1
selenium == 4.12.0
chromedriver_autoinstaller == 0.6.2
pyvirtualdisplay